tuya-lock-api/requirements.txt
fastapi
uvicorn
pycryptodome
tuya-iot-py-sdk
